Добавление ботов, поедание

+++ server.py
@@ -1,14 +1,12 @@
+import math
 import socket
-import time
-from os.path import split
-
-import psycopg2
 from sqlalchemy import create_engine
 from sqlalchemy.orm import declarative_base
 from sqlalchemy import Column, Integer, String
 from sqlalchemy.orm import sessionmaker
 import pygame
-
+from russian_names import RussianNames
+import random
 
 def find(vector: str):
     first = vector.find("<")
@@ -35,6 +33,12 @@ pygame.init()
 WIDTH_ROOM, HEIGHT_ROOM = 4000, 4000
 WIDTH_SERVER, HEIGHT_SERVER = 300, 300
 FPS = 100
+colors = ['Maroon', 'DarkRed', 'FireBrick', 'Red', 'Salmon', 'Tomato', 'Coral', 'OrangeRed', 'Chocolate', 'SandyBrown',
+        'DarkOrange', 'Orange', 'DarkGoldenrod', 'Goldenrod', 'Gold', 'Olive', 'Yellow', 'YellowGreen', 'GreenYellow',
+        'Chartreuse', 'LawnGreen', 'Green', 'Lime', 'SpringGreen', 'MediumSpringGreen', 'Turquoise',
+        'LightSeaGreen', 'MediumTurquoise', 'Teal', 'DarkCyan', 'Aqua', 'Cyan', 'DeepSkyBlue',
+        'DodgerBlue', 'RoyalBlue', 'Navy', 'DarkBlue', 'MediumBlue']
+MOBS_QUANTITY=25
 screen = pygame.display.set_mode((WIDTH_SERVER, HEIGHT_SERVER))
 pygame.display.set_caption('Сервер')
 clock = pygame.time.Clock()
@@ -89,8 +93,23 @@ class LocalPlayer:
         self.h_vision = 600
 
     def update(self):
-        self.x += self.speed_x
-        self.y += self.speed_y
+        if self.x -self.size<=0:
+            if self.speed_x>=0:
+                self.x+=self.speed_x
+        elif self.x+self.size>=WIDTH_ROOM:
+            if self.speed_x<=0:
+                self.x+=self.speed_x
+        else:
+            self.x+=self.speed_x
+
+        if self.y - self.size <= 0:
+            if self.speed_y >= 0:
+                self.y += self.speed_y
+        elif self.y + self.size >= HEIGHT_ROOM:
+            if self.speed_y <= 0:
+                self.y += self.speed_y
+        else:
+            self.y += self.speed_y
 
     def chnge_speed(self, vector):
         vector = find(vector)
@@ -134,35 +153,56 @@ main_socket.setblocking(False)
 main_socket.listen(5)
 print("Сокет создался")
 players = {}
+names=RussianNames(count=MOBS_QUANTITY*2, patronymic=False,surname=False, rare=True)
+names=list(set(names))
+for x in range(MOBS_QUANTITY):
+    server_mob=Players(names[x],None)
+    server_mob.abs_speed=2
+    server_mob.color=random.choice(colors)
+    server_mob.x, server_mob.y=random.randint(0,WIDTH_ROOM), random.randint(0,HEIGHT_ROOM)
+    server_mob.speed_x,server_mob.speed_y=random.randint(-1,1),random.randint(-1,1)
+    server_mob.size=random.randint(10,100)
+    s.add(server_mob)
+    s.commit()
+    local_mob=LocalPlayer(server_mob.id, server_mob.name, None, None).load()
+    players[server_mob.id] = local_mob
+tick=-1
 server_works = True
 while server_works:
     clock.tick(FPS)
-    try:
-        new_socket, addr = main_socket.accept()
-        new_socket.setblocking(False)
-        login = new_socket.recv(1024).decode()
-        print("Подключился", addr)
-        player = Players("Алла", addr)
-        if login.startswith("color"):
-            data = find_color(login[6:])
-            player.name, player.color = data
-        s.merge(player)
-        s.commit()
-        addr = f'({addr[0]},{addr[1]})'
-        data = s.query(Players).filter(Players.address == addr)
-        for user in data:
-            player = LocalPlayer(user.id, user.name, new_socket, addr).load()
-            players[user.id] = player
-    except BlockingIOError:
-        pass
-    for id in list(players):
+    tick+=1
+    if tick % 200 == 0:
         try:
-            data = players[id].sock.recv(1024).decode()
-            # print("Получил", data)
-            print(data)
-            players[id].chnge_speed(data)
-        except:
+            new_socket, addr = main_socket.accept()
+            new_socket.setblocking(False)
+            login = new_socket.recv(1024).decode()
+            print("Подключился", addr)
+            player = Players("Алла", addr)
+            if login.startswith("color"):
+                data = find_color(login[6:])
+                player.name, player.color = data
+            s.merge(player)
+            s.commit()
+            addr = f'({addr[0]},{addr[1]})'
+            data = s.query(Players).filter(Players.address == addr)
+            for user in data:
+                player = LocalPlayer(user.id, user.name, new_socket, addr).load()
+                players[user.id] = player
+        except BlockingIOError:
             pass
+    for id in list(players):
+        if players[id].sock is not None:
+            try:
+                data = players[id].sock.recv(1024).decode()
+                # print("Получил", data)
+                print(data)
+                players[id].chnge_speed(data)
+            except:
+                pass
+        else:
+            if tick%400==0:
+                vector=f"<{random.randint(-1,1)},{random.randint(-1,1)}>"
+                players[id].chnge_speed(vector)
     visible_bacteries={}
     for id in list(players):
         visible_bacteries[id]=[]
@@ -174,33 +214,50 @@ while server_works:
             dist_x=hero_2.x-hero_1.x
             dist_y=hero_2.y-hero_1.y
             if abs(dist_x) <= hero_1.w_vision//2+hero_2.size and abs(dist_y)<= hero_1.h_vision//2+hero_2.size:
-                x_=str(round(dist_x))
-                y_=str(round(dist_y))
-                size_=str(round(hero_2.size))
-                color_=hero_2.color
-                data=x_ +" "+y_+" "+size_+" "+color_
-                visible_bacteries[hero_1.id].append(data)
+                distance=math.sqrt(dist_x**2+dist_y**2)
+                if distance < hero_1.size and hero_1.size>1.1*hero_2.size:
+                    hero_2.size,hero_2.speed_x,hero_2.speed_y=0,0,0
+
+                if hero_1.addr is not None:
+                    x_=str(round(dist_x))
+                    y_=str(round(dist_y))
+                    size_=str(round(hero_2.size))
+                    color_=hero_2.color
+                    data=x_ +" "+y_+" "+size_+" "+color_
+                    visible_bacteries[hero_1.id].append(data)
 
             if abs(dist_x) <= hero_2.w_vision//2+hero_1.size and abs(dist_y)<= hero_2.h_vision//2+hero_1.size:
-                x_=str(round(-dist_x))
-                y_=str(round(-dist_y))
-                size_=str(round(hero_1.size))
-                color_=hero_1.color
-                data=x_ +" "+y_+" "+size_+" "+color_
-                visible_bacteries[hero_2.id].append(data)
+                distance=math.sqrt(dist_x**2+dist_y**2)
+                if distance < hero_2.size and hero_2.size>1.1*hero_1.size:
+                    hero_1.size,hero_1.speed_x,hero_1.speed_y=0,0,0
+                if hero_2.addr is not None:
+                    x_=str(round(-dist_x))
+                    y_=str(round(-dist_y))
+                    size_=str(round(hero_1.size))
+                    color_=hero_1.color
+                    data=x_ +" "+y_+" "+size_+" "+color_
+                    visible_bacteries[hero_2.id].append(data)
     for id in list(players):
         visible_bacteries[id]="<"+",".join(visible_bacteries[id])+">"
 
     for id in list(players):
-        try:
-            players[id].sock.send(visible_bacteries[id].encode())
-        except:
-            # players.remove(sock)
-            players[id].sock.close()
+        if players[id].sock is not None:
+            try:
+                players[id].sock.send(visible_bacteries[id].encode())
+            except:
+                # players.remove(sock)
+                players[id].sock.close()
+                del players[id]
+                s.query(Players).filter(Players.id == id).delete()
+                s.commit()
+                print("Сокет закрыт")
+    for id in list(players):
+        if players[id].errors>=500 or players[id].size==0:
+            if players[id].sock is not None:
+                players[id].sock.close()
             del players[id]
             s.query(Players).filter(Players.id == id).delete()
             s.commit()
-            print("Сокет закрыт")
     for event in pygame.event.get():
         if event.type == pygame.QUIT:
             server_works = False
@@ -211,13 +268,12 @@ while server_works:
         y = player.y * HEIGHT_SERVER // HEIGHT_ROOM
         size = player.size * WIDTH_SERVER // WIDTH_ROOM
         pygame.draw.circle(screen, player.color, (x, y), size)
-        s.query(Players).filter(Players.id == id).update({"size": size})
-        s.commit()
+
     for id in list(players):
         player = players[id]
         player.update()
-        s.query(Players).filter(Players.id == id).update({"x": player.x, "y": player.y})
-        s.commit()
+        # s.query(Players).filter(Players.id == id).update({"x": player.x, "y": player.y})
+        # s.commit()
 
     pygame.display.update()
 pygame.quit()
